Remove artifacts downloader from Provisioner (#817)
* remove artifacts downloader

* img bump
