+++ components/provisioner/internal/installation/release/gcs_downloader_test.go
@@ -201,3 +201,15 @@ func TestOnDemand_GetReleaseByVersion_Error(t *testing.T) {
 	}
 
 }
+
+type RoundTripFunc func(req *http.Request) *http.Response
+
+func (f RoundTripFunc) RoundTrip(req *http.Request) (*http.Response, error) {
+	return f(req), nil
+}
+
+func newTestClient(rtFunc RoundTripFunc) *http.Client {
+	return &http.Client{
+		Transport: RoundTripFunc(rtFunc),
+	}
+}
