--- components/provisioner/internal/installation/release/github_downloader.go
@@ -1,205 +1,0 @@
-package release
-
-import (
-	"context"
-	"encoding/json"
-	"errors"
-	"fmt"
-	"io/ioutil"
-	"net/http"
-	"sort"
-	"time"
-
-	"github.com/kyma-project/control-plane/components/provisioner/internal/util"
-
-	"github.com/kyma-project/control-plane/components/provisioner/internal/model"
-	"github.com/sirupsen/logrus"
-)
-
-const (
-	LongInterval  = 1 * time.Hour
-	ShortInterval = 1 * time.Minute
-
-	releaseFetchURL   = "https://api.github.com/repos/kyma-project/kyma/releases"
-	installerYAMLName = "kyma-installer-cluster.yaml"
-	tillerFormat      = "https://raw.githubusercontent.com/kyma-project/kyma/%s/installation/resources/tiller.yaml"
-)
-
-// Deprecated
-// Should be removed or switched to fetching from GCS bucket after version 1.14 is no longer supported
-func NewArtifactsDownloader(
-	repository Repository,
-	latestReleases int,
-	includePreReleases bool,
-	client *http.Client,
-	downloader TextFileDownloader,
-	log *logrus.Entry) *artifactsDownloader {
-	return &artifactsDownloader{
-		repository:         repository,
-		latestReleases:     latestReleases,
-		includePreReleases: includePreReleases,
-		httpClient:         client,
-		downloader:         downloader,
-		log:                log,
-	}
-}
-
-// Deprecated
-// Should be removed or switched to fetching from GCS bucket after version 1.14 is no longer supported
-type artifactsDownloader struct {
-	repository         Repository
-	latestReleases     int
-	includePreReleases bool
-	httpClient         *http.Client
-	downloader         TextFileDownloader
-	log                *logrus.Entry
-}
-
-func (ad artifactsDownloader) FetchPeriodically(ctx context.Context, shortInterval, longInterval time.Duration) {
-	for {
-		select {
-		case <-ctx.Done():
-			return
-		default:
-			err := ad.fetchLatestReleases()
-			if err != nil {
-				ad.log.Errorf("Error during release fetch: %s", err.Error())
-				time.Sleep(shortInterval)
-			} else {
-				time.Sleep(longInterval)
-			}
-		}
-	}
-}
-
-func (ad artifactsDownloader) fetchLatestReleases() error {
-	releases, err := ad.fetchReleases()
-	if err != nil {
-		return err
-	}
-
-	if !ad.includePreReleases {
-		releases = filterPreReleases(releases)
-	}
-
-	releases = getLatestReleases(releases, ad.latestReleases)
-
-	return ad.save(releases)
-}
-
-func (ad artifactsDownloader) fetchReleases() ([]model.GithubRelease, error) {
-	responseBody, err := ad.sendRequest(releaseFetchURL)
-	if err != nil {
-		return nil, err
-	}
-
-	var releases []model.GithubRelease
-	err = json.Unmarshal(responseBody, &releases)
-	if err != nil {
-		return nil, err
-	}
-
-	return releases, nil
-}
-
-func (ad artifactsDownloader) save(releases []model.GithubRelease) error {
-	for _, release := range releases {
-		artifacts, err := ad.buildRelease(release)
-		if err != nil {
-			return err
-		}
-
-		exists, err := ad.repository.ReleaseExists(artifacts.Version)
-
-		if err != nil {
-			return err
-		}
-
-		if !exists {
-			_, err = ad.repository.SaveRelease(artifacts)
-			if err != nil {
-				return err
-			}
-		}
-	}
-	return nil
-}
-
-func (ad artifactsDownloader) buildRelease(release model.GithubRelease) (model.Release, error) {
-	var installerURL string
-	for _, a := range release.Assets {
-		if a.Name == installerYAMLName {
-			installerURL = a.Url
-			break
-		}
-	}
-
-	if "" == installerURL {
-		return model.Release{}, errors.New(fmt.Sprintf("Release %d does not contain installer yaml", release.Id))
-	}
-
-	tillerURL := buildTillerURL(release.Name)
-
-	installerYAML, err := ad.downloader.Download(installerURL)
-	if err != nil {
-		return model.Release{}, err
-	}
-
-	tillerYAML, err := ad.downloader.DownloadOrEmpty(tillerURL)
-	if err != nil {
-		return model.Release{}, err
-	}
-
-	return model.Release{
-		Version:       release.Name,
-		TillerYAML:    tillerYAML,
-		InstallerYAML: installerYAML,
-	}, nil
-}
-
-func (ad artifactsDownloader) sendRequest(url string) ([]byte, error) {
-	resp, err := ad.httpClient.Get(url)
-	if err != nil {
-		return nil, err
-	}
-	defer util.Close(resp.Body)
-
-	if resp.StatusCode != http.StatusOK {
-		return nil, errors.New(fmt.Sprintf("Received unexpected http status %d", resp.StatusCode))
-	}
-
-	bytes, err := ioutil.ReadAll(resp.Body)
-	if err != nil {
-		return nil, err
-	}
-
-	return bytes, nil
-}
-
-func filterPreReleases(releases []model.GithubRelease) []model.GithubRelease {
-	var filtered []model.GithubRelease
-
-	for _, r := range releases {
-		if !r.Prerelease {
-			filtered = append(filtered, r)
-		}
-	}
-
-	return filtered
-}
-
-func getLatestReleases(releases []model.GithubRelease, latestReleases int) []model.GithubRelease {
-	sort.Slice(releases, func(i, j int) bool {
-		return releases[i].Id > releases[j].Id
-	})
-
-	if len(releases) < latestReleases {
-		return releases
-	}
-
-	return releases[:latestReleases]
-}
-
-func buildTillerURL(releaseName string) string {
-	return fmt.Sprintf(tillerFormat, releaseName)
-}
