+++ resources/kcp/values.yaml
@@ -9,7 +9,7 @@ global:
       version: "PR-808"
     provisioner:
       dir:
-      version: "PR-808"
+      version: "PR-817"
     kyma_environment_broker:
       dir:
       version: "PR-814"
